+++ test/unit/fuzz-tools.test.js
@@ -389,6 +389,7 @@ import * as idebergGlenoidV409 from '../../lib/ideberg-glenoid-v409.js';
 import * as andersonDalonzoV410 from '../../lib/anderson-dalonzo-v410.js';
 import * as levineEdwardsV411 from '../../lib/levine-edwards-v411.js';
 import * as lisfrancMyersonV412 from '../../lib/lisfranc-myerson-v412.js';
+import * as seinsheimerSubtrochV413 from '../../lib/seinsheimer-subtroch-v413.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -723,6 +724,7 @@ const MODULES = {
   'anderson-dalonzo-v410.js': andersonDalonzoV410,
   'levine-edwards-v411.js': levineEdwardsV411,
   'lisfranc-myerson-v412.js': lisfrancMyersonV412,
+  'seinsheimer-subtroch-v413.js': seinsheimerSubtrochV413,
   'tb-testing.js': tbTesting,
 };